Backups and restoring

+++ app/src/main/java/to/bitkit/ui/settings/SettingsScreen.kt
@@ -10,12 +10,17 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
+import org.lightningdevkit.ldknode.Network
 import to.bitkit.R
+import to.bitkit.env.Env
 import to.bitkit.ui.WalletViewModel
+import to.bitkit.ui.components.LabelText
 import to.bitkit.ui.components.NavButton
+import to.bitkit.ui.navigateToBackupSettings
 import to.bitkit.ui.navigateToDevSettings
 import to.bitkit.ui.navigateToGeneralSettings
 import to.bitkit.ui.navigateToLightning
+import to.bitkit.ui.navigateToRegtestSettings
 import to.bitkit.ui.scaffold.AppTopBar
 import to.bitkit.ui.scaffold.ScreenColumn
 
@@ -29,13 +34,18 @@ fun SettingsScreen(
         Column(
             verticalArrangement = Arrangement.spacedBy(8.dp),
             modifier = Modifier
-                .padding(horizontal = 24.dp)
+                .padding(horizontal = 16.dp)
                 .verticalScroll(rememberScrollState())
         ) {
             NavButton(stringResource(R.string.general)) { navController.navigateToGeneralSettings() }
+            NavButton(stringResource(R.string.button_backup_settings)) { navController.navigateToBackupSettings() }
             NavButton("Lightning") { navController.navigateToLightning() }
-            NavButton("Dev Settings") { navController.navigateToDevSettings() }
-            NavButton("Wipe Wallet", showIcon = false) { viewModel.wipeStorage() }
+            if (Env.network == Network.REGTEST) {
+                LabelText("REGTEST ONLY")
+                NavButton("Dev Settings") { navController.navigateToDevSettings() }
+                NavButton("Blocktank Regtest") { navController.navigateToRegtestSettings() }
+                NavButton("Wipe Wallet", showIcon = false) { viewModel.wipeStorage() }
+            }
         }
     }
 }
